requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
beautifulsoup4==4.12.3
python-whois==0.9.4
python-multipart==0.0.9
jinja2==3.1.4
markdown2==2.5.1
weasyprint==62.3